changed poll view settings
